allow multithreaded request serving
The server uses a single thread by deafult and tends to choke on multiple simultaneous requests. This change simply exposes werkzeug's http server threaded feature. The config.yaml can contain a threaded: true entry in the server section, and this allows the server to scale in a smoother fashion on multicore machines.